fix(hw_37): button state change
