+++ react/my-app/src/App.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-import { Wrapper } from './hw_37/wrapper';
+import { Wrapper } from './hw_37/Wrapper';
 
 function App() {
   return (
-    <Wrapper></Wrapper>   
+    <Wrapper></Wrapper>
   )
 }
