+++ react/my-app/src/hw_37/header.tsx
@@ -1,23 +1,10 @@
-import { ReactNode } from "react"
+import { ButtonMenu } from "./ButtonHeader"
 
 export const Header = () => (
     <header className="header">
-        <ButtonHamburger enabled handleClick={() => console.log("what's gonna here?!")}>click</ButtonHamburger>
+        <ButtonMenu defaultState handleClick={() => console.log('aaaaaaa')}/>
         <Title></Title>
     </header>
 )
 
 const Title = () => <div className="header__title">Sign In</div>
-
-type Menu = {
-    children?: ReactNode,
-    enabled: boolean,
-    handleClick: () => void
-}
-
-const ButtonHamburger = ({ children, enabled, handleClick }: Menu) => (
-    <button value={'123123'} className="header__button" onClick={handleClick} disabled={!enabled}>
-        {children}
-        {/* <span className="burger__line"></span> */}
-    </button>
-)
