+++ react/my-app/src/hw_37/wrapper.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react"
-import { Header } from "./header";
-import { Body } from "./body";
+import { Header } from "./Header";
+import { Body } from "./Body";
 
 type Props = { children?: ReactNode }
