+++ react/my-app/src/index.css
@@ -53,6 +53,8 @@ code {
   border: none;
   width: 150px;
   height: 40px;
+  font-size: 24px;
+  font-weight: 700;
   cursor: pointer;
   transition: 0.3s;
 }
